app/requirements.txt
streamlit==1.37.1
qdrant-client==1.9.1
sentence-transformers==3.0.1
torch==2.6.0
numpy==1.26.4
pandas==2.3.1
pyarrow==21.0.0
requests==2.32.3
python-dotenv==1.0.1
